adding to changes

+++ src/main/java/com/es/phoneshop/model/cart/CartItem.java
@@ -2,9 +2,9 @@ package com.es.phoneshop.model.cart;
 
 public class CartItem {
     private Long productId;
-    private Long quantity;
+    private int quantity;
 
-    public CartItem(Long productId, Long quantity) {
+    public CartItem(Long productId, int quantity) {
         if(productId == null) {
             throw new IllegalArgumentException("product is required");
         }
@@ -20,11 +20,11 @@ public class CartItem {
         this.productId = productId;
     }
 
-    public Long getQuantity() {
+    public int getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(Long quantity) {
+    public void setQuantity(int quantity) {
         this.quantity = quantity;
     }
